Fix typos in basic usage and change structure
I think it makes sense if we cover how to find deposits and collections
before explaning the deposit and collection landing pages. I have also
changed the page title to "navigating the data repository" to
diffirentiate from the "basic usage" section of the docs. I think this
is more clear for the prospective users.